chore: move plugins to top-level object
