+++ plugins/silly-effects.ts
@@ -1,8 +1,8 @@
 import type { Node, Schema } from "prosemirror-model";
 import { type EditorState, Plugin } from "prosemirror-state";
 import { DecorationSet, Decoration } from "prosemirror-view";
-import type { EditorPlugin, EditorPluginSpec } from "../../src";
-import type { MenuCommand } from "../../src/shared/menu";
+import type { EditorPlugin, EditorPluginSpec } from "../src";
+import type { MenuCommand } from "../src/shared/menu";
 
 const sillyEffectPlugin = new Plugin<{ decorations: DecorationSet }>({
     state: {
